contrast.bayes {arm}	R Documentation

Contrast Matrices

Description

Return a matrix of contrasts used in bayesglm.

Usage

contr.bayes.unordered(n, base = 1, contrasts = TRUE)
contr.bayes.ordered (n, scores = 1:n, contrasts = TRUE)

Arguments

n
a vector of levels for a factor, or the number of levels.

base
an integer specifying which group is considered the baseline group. Ignored if contrasts is FALSE.

contrasts
a logical indicating whether contrasts should be computed.

scores
the set of values over which orthogonal polynomials are to be computed.

Details

These functions are adapted from contr.treatment and contr.poly in stats package. The purpose for these functions are to keep the baseline levels of categorical variables and thus to suit the use of bayesglm.

contr.bayes.unordered is equivalent to contr.treatment whereas contr.bayes.ordered is equivalent to contr.poly.

Author(s)

Yu-Sung Su suyusung@tsinghua.edu.cn

See Also

C, contr.helmert, contr.poly, contr.sum, contr.treatment; glm, aov, lm, bayesglm.

Examples

 cat.var <- rep(1:3, 5)
 dim(contr.bayes.unordered(cat.var))
 # 15*15 baseline level kept!
 dim(contr.treatment(cat.var))
 # 15*14

[Package arm version 1.9-1 Index]