traceplot {arm}	R Documentation

Trace plot of ‘bugs’ object

Description

Displays a plot of iterations vs. sampled values for each variable in the chain, with a separate plot per variable.

Usage

## S4 method for signature 'bugs'
traceplot( x, mfrow = c( 1, 1 ), varname = NULL,
  match.head = TRUE, ask = TRUE,
  col = rainbow( x$n.chains ),
  lty = 1, lwd = 1, ...)

Arguments

x
A bugs object

mfrow
graphical parameter (see par)

varname
vector of variable names to plot

match.head
matches the variable names by the beginning of the variable names in bugs object

ask
logical; if TRUE, the user is asked before each plot, see par(ask=.).

col
graphical parameter (see par)

lty
graphical parameter (see par)

lwd
graphical parameter (see par)

...
further graphical parameters

Author(s)

Masanao Yajima yajima@stat.columbia.edu. Yu-Sung Su suyusung@tsinghua.edu.cn

See Also

densplot, plot.mcmc, traceplot

[Package arm version 1.9-1 Index]